Move types out to dedicated types directory, add documentation
