+++ src/types/LanguageFileExtensions.ts
@@ -1,0 +1,34 @@
+/**
+ * Associate a file extension with our normalized list of programming languages.
+ * Used to write the code block to file with a language-specific extension.
+ */
+export enum LanguageFileExtensions {
+  BASH = ".sh",
+  BATCH = ".bat",
+  C = ".c",
+  CFG = ".cfg",
+  CPP = ".cpp",
+  CSHARP = ".cs",
+  GO = ".go",
+  GROOVY = ".groovy",
+  INI = ".ini",
+  JAVA = ".java",
+  JAVASCRIPT = ".js",
+  JSON = ".json",
+  KOTLIN = ".kt",
+  NONE = ".txt",
+  PHP = ".php",
+  POWERSHELL = "powershell",
+  PYTHON = ".py",
+  RUBY = ".rb",
+  RUST = ".rs",
+  SCALA = ".scala",
+  SHELL = ".sh",
+  SQL = ".sql",
+  SWIFT = ".swift",
+  TEXT = ".txt",
+  TYPESCRIPT = ".ts",
+  UNDEFINED = ".txt",
+  XML = ".xml",
+  YAML = ".yaml",
+}
